// adjuster-portal-ui/src/app/models/files-client.model.ts
export class ClientFile {
  id?: string;
  client_id?: string;
  name: string;
  description?: string;
  type?: string;
  size?: number;
  path?: string;
  can_be_removed?: boolean = true;

  created_by: {
    first_name: string;
    last_name: string;
  }
  updated_by: {
    first_name: string;
    last_name: string;
  }
  created_at: Date;
  updated_at: Date;

}